backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlmodel==0.0.21
psycopg2-binary==2.9.9    # for Postgres; not used for SQLite
torch
torchvision
Pillow
email-validator
passlib==1.7.4
